Capture checkout intents through one shared test stub
Move stubCheckout, captureCheckoutIntent, and expectCapturedItemPriced
from server-reservation's local helpers into a new
test/test-utils/checkout.ts, and point every hand-rolled copy of the
same createCheckoutSession capture stub at it: the purchase and
book-parents locals, the renewal / site-assignment / qr-book /
custom-questions inline stubs, and the parents-gate wrapper.
