+++ test/lib/server-reservation/edge-cases.test.ts
@@ -8,10 +8,10 @@ import { getAttendeeOrderSummary } from "#shared/db/attendees/balance.ts";
 import { getDb } from "#shared/db/client.ts";
 import { modifiersTable } from "#shared/db/modifiers.ts";
 import { resetStripeClient, stripeApi } from "#shared/stripe.ts";
+import { captureCheckoutIntent } from "#test-utils/checkout.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
 import { mockRequest } from "#test-utils/mocks.ts";
 import {
-  captureCheckoutIntent,
   createProgrammeCharge,
   createSave10Promo,
   expectRefundedPlaceholder,
