+++ test/lib/server-reservation/promo-addons.test.ts
@@ -3,10 +3,10 @@ import { afterEach, it as test } from "@std/testing/bdd";
 import { stub } from "@std/testing/mock";
 import { handleRequest } from "#routes";
 import { resetStripeClient, stripeApi } from "#shared/stripe.ts";
+import { captureCheckoutIntent } from "#test-utils/checkout.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
 import { mockRequest } from "#test-utils/mocks.ts";
 import {
-  captureCheckoutIntent,
   createOptionalAddOn,
   createProgrammeCharge,
   createSave10Promo,
