+++ test/lib/server-reservation/public-default-modifiers.test.ts
@@ -6,6 +6,7 @@ import { modifierUsedQuantities } from "#shared/db/modifier-usage.ts";
 import { modifiersTable } from "#shared/db/modifiers.ts";
 import { resetStripeClient } from "#shared/stripe.ts";
 import { expectFlash } from "#test-utils/assertions.ts";
+import { captureCheckoutIntent } from "#test-utils/checkout.ts";
 import { submitTicketForm } from "#test-utils/csrf.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
 import { createTestListing } from "#test-utils/db-helpers/listings.ts";
@@ -13,7 +14,6 @@ import { setupStripe } from "#test-utils/settings.ts";
 import {
   addServiceCharge,
   attendeeCount,
-  captureCheckoutIntent,
   modifierUsageAmount,
   setPublicReservation,
   setupSoldOutModifierRace,
